Convert forceLightningComponentCreate
