+++ packages/salesforcedx-vscode-core/src/commands/index.ts
@@ -41,7 +41,9 @@ export { forceLightningAppCreate } from './forceLightningAppCreate';
 export {
   forceVisualforceComponentCreate
 } from './forceVisualforceComponentCreate';
-export { forceLightningComponentCreate } from './forceLightningComponentCreate';
+export {
+  forceLightningComponentCreate
+} from './templates/forceLightningComponentCreate';
 export { forceLightningEventCreate } from './forceLightningEventCreate';
 export { forceLightningInterfaceCreate } from './forceLightningInterfaceCreate';
 export { forceDebuggerStop } from './forceDebuggerStop';
