+++ packages/salesforcedx-vscode-core/src/commands/templates/forceLightningComponentCreate.ts
@@ -1,0 +1,58 @@
+/*
+ * Copyright (c) 2017, salesforce.com, inc.
+ * All rights reserved.
+ * Licensed under the BSD 3-Clause license.
+ * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
+ */
+
+import {
+  Command,
+  SfdxCommandBuilder
+} from '@salesforce/salesforcedx-utils-vscode/out/src/cli';
+import { DirFileNameSelection } from '@salesforce/salesforcedx-utils-vscode/out/src/types';
+import { nls } from '../../messages';
+import {
+  CompositeParametersGatherer,
+  LightningFilePathExistsChecker,
+  SelectFileName,
+  SelectOutputDir,
+  SfdxCommandlet,
+  SfdxWorkspaceChecker
+} from '../commands';
+import { BaseTemplateCommand } from './baseTemplateCommand';
+
+class ForceLightningComponentCreateExecutor extends BaseTemplateCommand {
+  public build(data: DirFileNameSelection): Command {
+    return new SfdxCommandBuilder()
+      .withDescription(nls.localize('force_lightning_component_create_text'))
+      .withArg('force:lightning:component:create')
+      .withFlag('--componentname', data.fileName)
+      .withFlag('--outputdir', data.outputdir)
+      .withLogName('force_lightning_component_create')
+      .build();
+  }
+
+  public createSubDirectory(): boolean {
+    return true;
+  }
+
+  public getFileExtension(): string {
+    return '.cmp';
+  }
+}
+
+const fileNameGatherer = new SelectFileName();
+const outputDirGatherer = new SelectOutputDir('aura', true);
+
+export async function forceLightningComponentCreate() {
+  const commandlet = new SfdxCommandlet(
+    new SfdxWorkspaceChecker(),
+    new CompositeParametersGatherer<DirFileNameSelection>(
+      fileNameGatherer,
+      outputDirGatherer
+    ),
+    new ForceLightningComponentCreateExecutor(),
+    new LightningFilePathExistsChecker()
+  );
+  await commandlet.run();
+}
